# openapi3_0 (3).json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Equipment registry API",
    "version": "1.0"
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:9090"
    }
  ],
...
  "paths": {
...
    "/refrigerators/lines": {
      "post": {
        "operationId": "addRefrigeratorLine",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "name"
                ],
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Название линейки"
                  },
                  "country": {
                    "type": "string",
                    "description": "Страна производитель"
                  },
                  "brand": {
                    "type": "string",
                    "description": "Фирма производитель"
                  },
                  "isAvailableOnline": {
                    "type": "boolean",
                    "description": "Возможность заказа онлайн(да/нет)"
                  },
                  "isInInstalment": {
                    "type": "boolean",
                    "description": "Возможность оформления рассрочки(да/нет)"
                  },
                  "models": {
                    "type": "array",
                    "description": "модели в наличии (выборка из моделей, которые представлены в виде справочника)",
                    "example": [
                      1,
                      2,
                      3
                    ],
                    "model": {
                      "type": "integer",
                      "description": "Идентификаторы моделей в наличии",
                      "format": "int64"
                    }
                  }
                },
                "description": "Данные для добавления новой линейки"
              }
            }
          },
          "description": "данные добавляемой линейки",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "name"
                  ],
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Название линейки"
                    },
                    "country": {
                      "type": "string",
                      "description": "Страна производитель"
                    },
                    "brand": {
                      "type": "string",
                      "description": "Фирма производитель"
                    },
                    "isAvailableOnline": {
                      "type": "boolean",
                      "description": "Возможность заказа онлайн(да/нет)"
                    },
                    "isInInstalment": {
                      "type": "boolean",
                      "description": "Возможность оформления рассрочки(да/нет)"
                    },
                    "models": {
                      "type": "array",
                      "description": "модели в наличии (выборка из моделей, которые представлены в виде справочника)",
                      "example": [
                        1,
                        2,
                        3
                      ],
                      "model": {
                        "type": "integer",
                        "description": "Идентификаторы моделей в наличии",
                        "format": "int64"
                      }
                    }
                  },
                  "description": "Данные для добавления новой линейки"
                }
              }
            },
            "description": "Линейка добавлена"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": "BAD_REQUEST",
                  "reason": "Incorrectly made request.",
                  "message": "Field: name. Error: must not be blank. Value: null",
                  "timestamp": "2022-09-07 09:10:50"
                }
              }
            },
            "description": "Запрос составлен некорректно"
          }
        },
        "summary": "Добавление новой линейки холодильников",
        "tags": [
          "Refrigerators"
        ]
      }
    },
...
    "/refrigerators/lines/{lineId}/models": {
      "post": {
        "operationId": "createRefrigeratorModel",
        "parameters": [
          {
            "description": "id линейки",
            "in": "path",
            "name": "lineId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "name"
                ],
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "наименование"
                  },
                  "colour": {
                    "type": "string",
                    "description": "цвет"
                  },
                  "size": {
                    "type": "string",
                    "description": "Размер"
                  },
                  "price": {
                    "type": "string",
                    "description": "Цена"
                  },
                  "doorCount": {
                    "type": "integer",
                    "description": "Количество дверей"
                  },
                  "compressor": {
                    "type": "string",
                    "description": "Компрессор"
                  },
                  "isAvailable": {
                    "type": "boolean",
                    "description": "Наличие товара"
                  },
                  "type": {
                    "type": "integer",
                    "description": "Идентификатор линейки"
                  }
                },
                "description": "Модель холодильника"
              }
            }
          },
          "description": "данные новой модели",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "name"
                  ],
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "наименование"
                    },
                    "colour": {
                      "type": "string",
                      "description": "цвет"
                    },
                    "size": {
                      "type": "string",
                      "description": "Размер"
                    },
                    "price": {
                      "type": "string",
                      "description": "Цена"
                    },
                    "doorCount": {
                      "type": "integer",
                      "description": "Количество дверей"
                    },
                    "compressor": {
                      "type": "string",
                      "description": "Компрессор"
                    },
                    "isAvailable": {
                      "type": "boolean",
                      "description": "Наличие товара"
                    },
                    "type": {
                      "type": "integer",
                      "description": "Идентификатор линейки"
                    }
                  },
                  "description": "Модель холодильника"
                }
              }
            },
            "description": "Модель добавлена"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": "BAD_REQUEST",
                  "reason": "Incorrectly made request.",
                  "message": "Field: title. Error: must not be blank. Value: null",
                  "timestamp": "2022-09-07 09:10:50"
                }
              }
            },
            "description": "Запрос составлен некорректно"
          }
        },
        "summary": "Добавление модели в линейку холодильников",
        "tags": [
          "Refrigerators"
        ]
      }
    },
...
    "/refrigerators/models/sorted": {
      "get": {
        "description": "Эндпоинт возвращает сортировку холодильников по алфавиту и по стоимости",
        "operationId": "getRefrigeratorsSortedByPriceAndAlphabet",
        "parameters": [
          {
            "description": "цена по возрастанию - asc, по убыванию - desc",
            "in": "query",
            "name": "price",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "количество моделей, которые нужно пропустить для формирования текущего набора",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "description": "количество моделей в наборе",
            "in": "query",
            "name": "sizePage",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "models": {
                    "$ref": "#/components/schemas/RefrigeratorModel"
                  }
                }
              }
            },
            "description": "Модели отсортированы по алфавиту и по цене"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": "BAD_REQUEST",
                  "reason": "Incorrectly made request.",
                  "message": "Failed to convert value of type java.lang.String to required type int; nested exception is java.lang.NumberFormatException: For input string: ad",
                  "timestamp": "2022-09-07 09:10:50"
                }
              }
            },
            "description": "Запрос составлен некорректно"
          }
        },
        "summary": "Сортировка моделей холодильников",
        "tags": [
          "Refrigerators"
        ]
      }
    },
...
    "/smartphones/models/filtered": {
      "get": {
        "description": "Поиск (по наименованию) должен быть без учета регистра букв.\\n В случае, если по заданным фильтрам не найдено ни одной модели, возвращает пустой список",
        "operationId": "findByNameAndFilterByParams",
        "parameters": [
          {
            "description": "текст для поиска в наименовании",
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "цвет",
            "in": "query",
            "name": "colour",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ценa от",
            "in": "query",
            "name": "priceFrom",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ценa до",
            "in": "query",
            "name": "priceTo",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "память",
            "in": "query",
            "name": "memory",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "камера",
            "in": "query",
            "name": "camera",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "количество моделей, которые нужно пропустить для формирования текущего набора",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "description": "количество моделей  в наборе",
            "in": "query",
            "name": "sizePage",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "models": {
                    "$ref": "#/components/schemas/SmartphoneModel"
                  }
                }
              }
            },
            "description": "Модели найдены"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": "BAD_REQUEST",
                  "reason": "Incorrectly made request.",
                  "message": "Event must be published",
                  "timestamp": "2022-09-07 09:10:50"
                }
              }
            },
            "description": "Запрос составлен некорректно"
          }
        },
        "summary": "Получение телефонов с возможностью фильтрации",
        "tags": [
          "Smartphones"
        ]
      }
    },
...
    "/refrigerators/models/filtered": {
      "get": {
        "description": "Поиск (по наименованию) должен быть без учета регистра букв.\\n В случае, если по заданным фильтрам не найдено ни одной модели, возвращает пустой список",
        "operationId": "findRefrigeratorsByNameAndFilterByParams",
        "parameters": [
          {
            "description": "текст для поиска в наименовании",
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "цвет",
            "in": "query",
            "name": "colour",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ценa от",
            "in": "query",
            "name": "priceFrom",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ценa до",
            "in": "query",
            "name": "priceTo",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "количество дверей",
            "in": "query",
            "name": "doorCount",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "компрессор",
            "in": "query",
            "name": "compressor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "количество моделей, которые нужно пропустить для формирования текущего набора",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "description": "количество моделей в наборе",
            "in": "query",
            "name": "sizePage",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "models": {
                    "$ref": "#/components/schemas/RefrigeratorModel"
                  }
                }
              }
            },
            "description": "Модели найдены"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": "BAD_REQUEST",
                  "reason": "Incorrectly made request.",
                  "message": "Event must be published",
                  "timestamp": "2022-09-07 09:10:50"
                }
              }
            },
            "description": "Запрос составлен некорректно"
          }
        },
        "summary": "Получение холодильников с возможностью фильтрации",
        "tags": [
          "Refrigerators"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SmartphoneModel": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "наименование"
          },
          "colour": {
            "type": "string",
            "description": "цвет"
          },
          "size": {
            "type": "string",
            "description": "Размер"
          },
          "price": {
            "type": "string",
            "description": "Цена"
          },
          "isAvailable": {
            "type": "boolean",
            "description": "Наличие товара"
          },
          "memory": {
            "type": "integer",
            "description": "Память"
          },
          "cameraCount": {
            "type": "integer",
            "description": "Камера"
          },
          "type": {
            "type": "integer",
            "description": "Идентификатор линейки"
          }
        },
        "description": "Модель телефона"
      },
...
      "RefrigeratorModel": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "наименование"
          },
          "colour": {
            "type": "string",
            "description": "цвет"
          },
          "size": {
            "type": "string",
            "description": "Размер"
          },
          "price": {
            "type": "string",
            "description": "Цена"
          },
          "doorCount": {
            "type": "integer",
            "description": "Количество дверей"
          },
          "compressor": {
            "type": "string",
            "description": "Компрессор"
          },
          "isAvailable": {
            "type": "boolean",
            "description": "Наличие товара"
          },
          "type": {
            "type": "integer",
            "description": "Идентификатор линейки"
          }
        },
        "description": "Модель холодильника"
      },
...
    }
  }
}